Implemented SLTIU And did MAJOR change to  How REgisters are selected
Made the input of the Reg File B select take IR[20:16] and modified all the MUXPA selects
So A in ALU always has the corresponding RS and B has either RT or immediate depending
on the instruction.
NOTE: It is possible that we do not need the new Select in MUXPA Since we usually either need
IR[25:21] or Reg. 0. Keep eye out for any new instructions that may
use IR[20:16] in MuxPA